create-and-test-generative-model-metric/requirements.txt
torch
numpy
Pillow
tqdm
transformers
accelerate
absl-py
streamlit
timm